refactor: 💡 move WorkerResponse to /worker folder
