+++ src/worker/WorkerResponse.ts
@@ -1,6 +1,5 @@
-import type {TransferList} from './types';
+import type {TransferList} from '../types';
 
-// TODO: Move this to /worker
 export class WorkerResponse<D = unknown> {
   constructor(public readonly data: D, public readonly transferList: TransferList) {}
 }
