+++ src/worker/types.ts
@@ -1,4 +1,4 @@
-import type {WorkerResponse} from '../WorkerResponse';
+import type {WorkerResponse} from './WorkerResponse';
 import type {WpSend, WpRecv} from '../types';
 
 export interface WorkerModule {
